requirements.txt
python-dotenv
pyyaml
aiohttp